requirements.txt
langserve
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
futures

langchain == 0.3
langchain_community == 0.3
langchain_openai ==0.2
langchain-ollama ==0.2
langchain-groq == 0.2
langchain-core == 0.3.0
langchain-text-splitters == 0.3
langchain-experimental ==  0.3
pydantic ==2.7.4
requests ==2.28
aiohttp == 3.8.3
tenacity == 8.1.0
jsonpatch == 1.33